implement authorizaton checks on the backend

+++ app/Http/Controllers/RoleController.php
@@ -2,19 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\DestroyRoleRequest;
 use App\Models\Role;
 use App\Tables\Roles;
 use App\Models\Permission;
 use Illuminate\Http\Request;
+use ProtoneMedia\Splade\Facades\Toast;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
-use ProtoneMedia\Splade\Facades\Toast;
+use App\Http\Requests\DestroyRoleRequest;
 
 class RoleController extends Controller
 {
     public function index()
     {
+        $this->authorize('role_access');
+
         return view('roles.index', [
             'roles' => Roles::class,
         ]);
@@ -22,6 +24,8 @@ class RoleController extends Controller
 
     public function create()
     {
+        $this->authorize('role_create');
+
         $permissions = Permission::all();
 
         return view('roles.create', compact('permissions'));
@@ -29,6 +33,8 @@ class RoleController extends Controller
 
     public function store(StoreRoleRequest $request)
     {
+        $this->authorize('role_create');
+
         $role = Role::create([
             'title' => $request->title,
         ]);
@@ -40,6 +46,8 @@ class RoleController extends Controller
 
     public function edit(Role $role)
     {
+        $this->authorize('role_edit');
+
         $permissions = Permission::all();
 
         return view('roles.edit', compact('permissions', 'role'));
@@ -47,6 +55,8 @@ class RoleController extends Controller
 
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        $this->authorize('role_edit');
+
         $role->update([
             'title' => $request->title,
         ]);
@@ -60,6 +70,8 @@ class RoleController extends Controller
 
     public function destroy(DestroyRoleRequest $request, Role $role)
     {
+        $this->authorize('role_delete');
+
         $role->delete();
 
         Toast::title('Role deleted successfully.');
